Updated project files

+++ server.js
@@ -1,6 +1,7 @@
 // server.js
 const net = require('net');
 const fs = require('fs');
+const path = require('path');
 const port = 8124;
 const firstRequestStr = 'FILES';
 const successReq = 'ASC';
@@ -14,15 +15,30 @@ const server = net.createServer((client) => {
 
   client.setEncoding('utf8');
   client.ID = Date.now() + seed++;
+  client.RequestNumber = 0;
 
   client.on('data', (data) => {
-    console.log(data);
 
-    if (data == firstRequestStr){
+    client.RequestNumber = client.RequestNumber + 1;
+
+    if ( (data == firstRequestStr) && (client.RequestNumber == 1) ){
+      console.log(data);
       client.write(successReq);
-    }else{
+    }else if ( (data != firstRequestStr) && (client.RequestNumber == 1) ){
+      console.log(data);
       client.write(failedReq);
       client.destroy();
+    }else{
+      console.log(data);
+
+      fs.mkdir(`${client.ID}`, { recursive: true }, (err) => {
+          if (err) throw err;
+
+          fs.writeFile(path.join(serverFiles, `${client.ID}`, data), data, (err) => {
+            if (err) throw err;
+            console.log('The file has been saved!');
+          });
+      });
     }
   });
